project_django/requirements.txt
asgiref==3.2.3
astroid==2.3.3
authlib==0.13
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
django-cors-headers==3.2.0
django-filter==2.2.0
django-reversion==3.0.5
django-six==1.0.4
django-widget-tweaks==1.4.5
django==3.0.1
djangorestframework==3.11.0
ecdsa==0.14.1
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itypes==1.1.0
jinja2==2.10.3
keycloak==2.0.0
lazy-object-proxy==1.4.3
markdown==3.1.1
markupsafe==1.1.1
mccabe==0.6.1
numpy==1.18.0
oauthlib==3.1.0
pandas==0.25.3
pillow==6.2.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycparser==2.19
pygments==2.5.2
pyjwt==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-jose==3.1.0
python-keycloak-client==0.2.2
python-keycloak==0.18.0
python-memcached==1.59
pytz==2019.3
pyyaml==5.2
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.0
six==1.13.0
sqlalchemy==1.3.12
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
wrapt==1.11.2
xlrd==1.2.0